--- lightning_pod/cli/utils/build.py
@@ -1,9 +1,0 @@
-from lightning_pod.utils.new.new_project import (
-    preserve_examples,
-    make_new_lightning_pod,
-)
-
-
-def main():
-    preserve_examples()
-    make_new_lightning_pod()
